Join Room Role Separation
- Added specification for if you want to play or watch when you join
room
- Added alias redirect routes to easily link someone to join as player
  or as audience member
