+++ assets/js/components/pages/room/index.tsx
@@ -1,0 +1,2 @@
+export { RoomPageContent } from "./RoomPageContent";
+export { RoomPageJoin } from "./RoomPageJoin";
