+++ assets/js/utils/transforms.spec.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { loopToEvents, midiToPitch } from "./";
 import { Loop, Note, TimestepSlice, LocalNoteEvent } from "../types";
+import { INPUT_LAG_COMPENSATION } from "../constants";
 
 const loopMocks = require("../mocks/loops.json");
 
@@ -10,10 +11,20 @@ describe("loopToEvents", () => {
     const now = 0;
     const timestepSize = 1_000_000;
     const expectedEvents = [
-      { time: now, note: midiToPitch(72), velocity: 1, duration: 1 },
-      { time: 8, note: midiToPitch(76), velocity: 0, duration: 2 },
       {
-        time: 12,
+        time: now + INPUT_LAG_COMPENSATION,
+        note: midiToPitch(72),
+        velocity: 1,
+        duration: 1,
+      },
+      {
+        time: 8 + INPUT_LAG_COMPENSATION,
+        note: midiToPitch(76),
+        velocity: 0,
+        duration: 2,
+      },
+      {
+        time: 12 + INPUT_LAG_COMPENSATION,
         note: midiToPitch(79),
         velocity: 0.7480314960629921,
         duration: 1,
